Made changes to ITsupport and index.html
Created separate views for IT employees and users. TODO: Find a way to "hide" itHome or userHome based on what the person's role is. I added the .hide and .show in the script file, but it doesn't look like it has any effect. Not sure why it works for homeComponent, but not for the others.